+++ tests/unit/mixin/test_semantics.py
@@ -3,7 +3,6 @@ from pathlib import Path
 
 from pyiron_workflow.mixin.semantics import (
     CyclicPathError,
-    ParentMost,
     Semantic,
     SemanticParent,
 )
@@ -15,15 +14,9 @@ class ConcreteParent(SemanticParent[Semantic]):
         return Semantic
 
 
-class ConcreteParentMost(ParentMost[Semantic]):
-    @classmethod
-    def child_type(cls) -> type[Semantic]:
-        return Semantic
-
-
 class TestSemantics(unittest.TestCase):
     def setUp(self):
-        self.root = ConcreteParentMost("root")
+        self.root = ConcreteParent("root")
         self.child1 = Semantic("child1", parent=self.root)
         self.middle1 = ConcreteParent("middle", parent=self.root)
         self.middle2 = ConcreteParent("middle_sub", parent=self.middle1)
@@ -69,18 +62,6 @@ class TestSemantics(unittest.TestCase):
         with self.subTest("Normal usage"):
             self.assertEqual(self.child1.parent, self.root)
             self.assertEqual(self.root.parent, None)
-
-        with self.subTest(f"{ParentMost.__name__} exceptions"):
-            with self.assertRaises(
-                TypeError, msg=f"{ParentMost.__name__} instances can't have parent"
-            ):
-                self.root.parent = ConcreteParent(label="foo")
-
-            with self.assertRaises(
-                TypeError, msg=f"{ParentMost.__name__} instances can't be children"
-            ):
-                some_parent = ConcreteParent(label="bar")
-                some_parent.add_child(self.root)
 
         with self.subTest("Cyclicity exceptions"):
             with self.assertRaises(CyclicPathError):
